Split markdown.tsx into several different files.
