+++ src/renderer/src/model/markdown/markdown.tsx
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+import { renderAST } from "./renderer";
+import { tokenize } from "./tokenizer";
+import { parse } from "./parser";
+
+
+export function renderMarkdown(markdownString: string): ReactNode[] {
+  return renderAST(parse(tokenize(markdownString)));
+}
