+++ src/renderer/src/model/markdown/token.type.ts
@@ -1,0 +1,32 @@
+export type ListPoint = 
+  "list-main" | 
+  "list-info" | 
+  "list-conclusion" | 
+  "list-question" | 
+  "list-important" | 
+  "list-pro" | 
+  "list-con"
+;
+
+export type TokenType = 
+  "strong" | 
+  "emphasized" | 
+  "paragraph" | 
+  "header" | 
+  "plain-text" | 
+  "new-line" | 
+  "tab-character" | 
+  "ignore-next" | 
+  "white-space" | 
+  "link-label-open" |
+  "link-label-close" |
+  "link-open" |
+  "link-close" |
+  "link" | 
+  "chart-open" |
+  "chart-close" |
+  "chart" | 
+  "underline-open" | 
+  "underline-close" |
+  "underlined" |
+  ListPoint;
